TNO-1751 PR updates #3
