+++ app/editor/src/features/content/list-view/utils/index.ts
@@ -1,5 +1,6 @@
 export * from './applySortBy';
 export * from './getActions';
+export * from './getFilter';
 export * from './getStatusText';
 export * from './makeFilter';
 export * from './queryToFilter';
